ticket:1086 - added conditional compilation for common
Summary: common is compiled conditionally for when rasnet is enabled/disabled. --enable-strict is working in both cases.

Test Plan: manually

Reviewers: atoader, dmisev, mdumitru

Reviewed By: dmisev, mdumitru

Subscribers: vmerticariu

Differential Revision: http://codereview.rasdaman.org/D62

+++ common/test/time/timertest.cc
@@ -51,14 +51,14 @@ TEST_F(TimerTest, CheckIfTimerIsNotExpiredWhenTimeHasNotPassed)
 TEST_F(TimerTest, SleepUntilTimerExpires)
 {
     //Test if the timer has expired
-    usleep(period * 1000);
+    usleep((uint32_t)period * 1000u);
     ASSERT_TRUE(timer.hasExpired());
 }
 
 TEST_F(TimerTest, SleepUntilTimerExpiresAndReset)
 {
     //Test if the timer has expired
-    usleep(period * 1000);
+    usleep((uint32_t)period * 1000u);
     timer.reset();
     ASSERT_FALSE(timer.hasExpired());
 }
